refactor(riscv): colocate loadstore width tests
