+++ extensions/riscv/circuit/src/load_sign_extend/mod.rs
@@ -8,4 +8,6 @@ pub use halfword::*;
 pub use word::*;
 
 #[cfg(test)]
+pub(crate) mod test_utils;
+#[cfg(test)]
 mod tests;
