+++ extensions/riscv/circuit/src/load_sign_extend/test_utils.rs
@@ -1,6 +1,6 @@
-use std::{array, borrow::BorrowMut, sync::Arc};
+pub(crate) use std::{array, borrow::BorrowMut, sync::Arc};
 
-use openvm_circuit::{
+pub(crate) use openvm_circuit::{
     arch::{
         testing::{
             memory::gen_pointer, TestBuilder, TestChipHarness, VmChipTestBuilder,
@@ -10,13 +10,15 @@ use openvm_circuit::{
     },
     system::memory::merkle::public_values::PUBLIC_VALUES_AS,
 };
-use openvm_circuit_primitives::bitwise_op_lookup::{
+pub(crate) use openvm_circuit_primitives::bitwise_op_lookup::{
     BitwiseOperationLookupAir, BitwiseOperationLookupBus, BitwiseOperationLookupChip,
     SharedBitwiseOperationLookupChip,
 };
-use openvm_instructions::{instruction::Instruction, riscv::RV64_REGISTER_AS, LocalOpcode};
-use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{self, *};
-use openvm_stark_backend::{
+pub(crate) use openvm_instructions::{
+    instruction::Instruction, riscv::RV64_REGISTER_AS, LocalOpcode,
+};
+pub(crate) use openvm_riscv_transpiler::Rv64LoadStoreOpcode::{self, *};
+pub(crate) use openvm_stark_backend::{
     p3_air::BaseAir,
     p3_field::PrimeCharacteristicRing,
     p3_matrix::{
@@ -25,10 +27,10 @@ use openvm_stark_backend::{
     },
     utils::disable_debug_builder,
 };
-use openvm_stark_sdk::{p3_baby_bear::BabyBear, utils::create_seeded_rng};
-use rand::{rngs::StdRng, Rng};
+pub(crate) use openvm_stark_sdk::{p3_baby_bear::BabyBear, utils::create_seeded_rng};
+pub(crate) use rand::{rngs::StdRng, Rng};
 #[cfg(feature = "cuda")]
-use {
+pub(crate) use {
     super::{
         Rv64LoadSignExtendByteChipGpu, Rv64LoadSignExtendHalfwordChipGpu,
         Rv64LoadSignExtendWordChipGpu,
@@ -44,7 +46,7 @@ use {
     openvm_circuit_primitives::var_range::VariableRangeCheckerChip,
 };
 
-use super::{
+pub(crate) use super::{
     byte::{
         LoadSignExtendByteCoreAir, LoadSignExtendByteCoreCols, LoadSignExtendByteFiller,
         Rv64LoadSignExtendByteAir, Rv64LoadSignExtendByteChip, Rv64LoadSignExtendByteExecutor,
@@ -60,7 +62,7 @@ use super::{
         LOAD_SIGN_EXTEND_WORD_SELECTOR_WIDTH,
     },
 };
-use crate::{
+pub(crate) use crate::{
     adapters::{
         rv64_bytes_to_u16_block, rv64_bytes_to_u32, rv64_u16_block_to_bytes, sign_extend_imm16,
         Rv64LoadStoreAdapterAir, Rv64LoadStoreAdapterExecutor, Rv64LoadStoreAdapterFiller,
@@ -70,34 +72,30 @@ use crate::{
     loadstore::common::run_write_data,
 };
 
-const IMM_BITS: usize = 16;
-const MAX_INS_CAPACITY: usize = 128;
-type F = BabyBear;
+pub(crate) const IMM_BITS: usize = 16;
+pub(crate) const MAX_INS_CAPACITY: usize = 128;
+pub(crate) type F = BabyBear;
 
-mod byte;
-mod halfword;
-mod word;
-
-type ByteHarness = TestChipHarness<
+pub(crate) type ByteHarness = TestChipHarness<
     F,
     Rv64LoadSignExtendByteExecutor,
     Rv64LoadSignExtendByteAir,
     Rv64LoadSignExtendByteChip<F>,
 >;
-type HalfwordHarness = TestChipHarness<
+pub(crate) type HalfwordHarness = TestChipHarness<
     F,
     Rv64LoadSignExtendHalfwordExecutor,
     Rv64LoadSignExtendHalfwordAir,
     Rv64LoadSignExtendHalfwordChip<F>,
 >;
-type WordHarness = TestChipHarness<
+pub(crate) type WordHarness = TestChipHarness<
     F,
     Rv64LoadSignExtendWordExecutor,
     Rv64LoadSignExtendWordAir,
     Rv64LoadSignExtendWordChip<F>,
 >;
 
-fn create_byte_harness(
+pub(crate) fn create_byte_harness(
     tester: &mut VmChipTestBuilder<F>,
 ) -> (
     ByteHarness,
@@ -143,7 +141,7 @@ fn create_byte_harness(
     )
 }
 
-fn create_halfword_harness(tester: &mut VmChipTestBuilder<F>) -> HalfwordHarness {
+pub(crate) fn create_halfword_harness(tester: &mut VmChipTestBuilder<F>) -> HalfwordHarness {
     let range_checker = tester.range_checker();
     let air = Rv64LoadSignExtendHalfwordAir::new(
         Rv64LoadStoreAdapterAir::new(
@@ -169,7 +167,7 @@ fn create_halfword_harness(tester: &mut VmChipTestBuilder<F>) -> HalfwordHarness
     HalfwordHarness::with_capacity(executor, air, chip, MAX_INS_CAPACITY)
 }
 
-fn create_word_harness(tester: &mut VmChipTestBuilder<F>) -> WordHarness {
+pub(crate) fn create_word_harness(tester: &mut VmChipTestBuilder<F>) -> WordHarness {
     let range_checker = tester.range_checker();
     let air = Rv64LoadSignExtendWordAir::new(
         Rv64LoadStoreAdapterAir::new(
@@ -196,7 +194,7 @@ fn create_word_harness(tester: &mut VmChipTestBuilder<F>) -> WordHarness {
 }
 
 #[allow(clippy::too_many_arguments)]
-fn set_and_execute<RA: Arena, E: PreflightExecutor<F, RA>>(
+pub(crate) fn set_and_execute<RA: Arena, E: PreflightExecutor<F, RA>>(
     tester: &mut impl TestBuilder<F>,
     executor: &mut E,
     arena: &mut RA,
@@ -276,13 +274,13 @@ fn set_and_execute<RA: Arena, E: PreflightExecutor<F, RA>>(
     }
 }
 
-fn memory_config_for() -> MemoryConfig {
+pub(crate) fn memory_config_for() -> MemoryConfig {
     let mut mem_config = MemoryConfig::default();
     mem_config.addr_spaces[RV64_REGISTER_AS as usize].num_cells = 1 << 29;
     mem_config.addr_spaces[PUBLIC_VALUES_AS as usize].num_cells = 1 << 29;
     mem_config
 }
-fn assert_pranked_byte_fails(prank: impl Fn(&mut LoadSignExtendByteCoreCols<F>)) {
+pub(crate) fn assert_pranked_byte_fails(prank: impl Fn(&mut LoadSignExtendByteCoreCols<F>)) {
     let mut rng = create_seeded_rng();
     let mut tester = VmChipTestBuilder::from_config(memory_config_for());
     let (mut harness, bitwise) = create_byte_harness(&mut tester);
@@ -313,7 +311,7 @@ fn assert_pranked_byte_fails(prank: impl Fn(&mut LoadSignExtendByteCoreCols<F>))
         .expect_err("pranked signed byte load trace should fail");
 }
 
-fn assert_pranked_halfword_fails(
+pub(crate) fn assert_pranked_halfword_fails(
     prank: impl Fn(&mut LoadSignExtendAlignedCoreCols<F, LOAD_SIGN_EXTEND_HALFWORD_SELECTOR_WIDTH>),
 ) {
     let mut rng = create_seeded_rng();
@@ -345,7 +343,7 @@ fn assert_pranked_halfword_fails(
         .expect_err("pranked signed halfword load trace should fail");
 }
 
-fn assert_pranked_word_fails(
+pub(crate) fn assert_pranked_word_fails(
     prank: impl Fn(&mut LoadSignExtendAlignedCoreCols<F, LOAD_SIGN_EXTEND_WORD_SELECTOR_WIDTH>),
 ) {
     let mut rng = create_seeded_rng();
@@ -377,26 +375,21 @@ fn assert_pranked_word_fails(
         .expect_err("pranked signed word load trace should fail");
 }
 
-#[test]
-fn negative_split_signed_load_tests() {
-    assert_pranked_byte_fails(|core| core.data_most_sig_bit += F::ONE);
-    assert_pranked_halfword_fails(|core| core.data_most_sig_bit += F::ONE);
-    assert_pranked_word_fails(|core| core.data_most_sig_bit += F::ONE);
-}
-
 // ////////////////////////////////////////////////////////////////////////////////////
 //  CUDA TESTS
 //
 //  Ensure GPU tracegen is equivalent to CPU tracegen.
 // ////////////////////////////////////////////////////////////////////////////////////
 #[cfg(feature = "cuda")]
-fn dummy_range_checker() -> Arc<VariableRangeCheckerChip> {
+pub(crate) fn dummy_range_checker() -> Arc<VariableRangeCheckerChip> {
     Arc::new(VariableRangeCheckerChip::new(
         default_var_range_checker_bus(),
     ))
 }
 #[cfg(feature = "cuda")]
-fn transfer_load_sign_extend_records<G, C, A, E>(harness: &mut GpuTestChipHarness<F, E, A, G, C>) {
+pub(crate) fn transfer_load_sign_extend_records<G, C, A, E>(
+    harness: &mut GpuTestChipHarness<F, E, A, G, C>,
+) {
     type Record<'a> = (&'a mut Rv64LoadStoreAdapterRecord, &'a mut LoadStoreRecord);
     harness
         .dense_arena
