+++ extensions/riscv/circuit/src/loadstore/byte/mod.rs
@@ -13,6 +13,9 @@ mod cuda;
 #[cfg(feature = "cuda")]
 pub use cuda::*;
 
+#[cfg(test)]
+mod tests;
+
 pub type Rv64LoadStoreByteAir = VmAirWrapper<Rv64LoadStoreAdapterAir, LoadStoreByteCoreAir>;
 pub type Rv64LoadStoreByteExecutor = LoadStoreExecutor<Rv64LoadStoreAdapterExecutor, KIND_BYTE>;
 pub type Rv64LoadStoreByteChip<F> = VmChipWrapper<F, LoadStoreByteFiller>;
