+++ extensions/riscv/circuit/src/loadstore/halfword/mod.rs
@@ -29,3 +29,6 @@ pub type Rv64LoadStoreHalfwordChip<F> = VmChipWrapper<F, LoadStoreHalfwordFiller
 mod cuda;
 #[cfg(feature = "cuda")]
 pub use cuda::*;
+
+#[cfg(test)]
+mod tests;
