+++ extensions/riscv/circuit/src/loadstore/word/mod.rs
@@ -27,3 +27,6 @@ pub type Rv64LoadStoreWordChip<F> = VmChipWrapper<F, LoadStoreWordFiller>;
 mod cuda;
 #[cfg(feature = "cuda")]
 pub use cuda::*;
+
+#[cfg(test)]
+mod tests;
