+++ extensions/riscv/circuit/src/loadstore/word/tests.rs
@@ -1,6 +1,6 @@
 use test_case::test_case;
 
-use super::*;
+use crate::loadstore::test_utils::*;
 
 #[test]
 fn positive_loadwu_shift4_test() {
